requirements.txt
groq
crewai
langchain-groq
crewai.tools
streamlit
python-dotenv